Validate parsed Perana Volleyball scouting data

Validate parsed Perana Volleyball scouting data

pv_validate(x, validation_level = 2)

Arguments

x
peranavolley: peranavolley object as returned by pv_read

validation_level
numeric: how strictly to check? If 0, perform no checking; if 1, only identify major errors; if 2, also return any issues that are likely to lead to misinterpretation of data; if 3, return all issues (including minor issues)

Value

data.frame with columns message (the validation message), file_line_number (the corresponding line number in the file), video_time, and file_line (the actual line from the file).

See also

Examples

filename <- pv_example_file()
x <- pv_read(filename)
pv_validate(x)

#>    file_line_number video_time
#> 3               237          0
#> 4               820          0
#> 5               821          0
#> 21              310          0
#> 22              186          0
#> 23              198          0
#> 24              228          0
#> 25              316          0
#> 26              461          0
#> 27              631          0
#> 28              635          0
#> 29              757          0
#> 30              816          0
#> 31              264          0
#> 32              824          0
#> 33              321          0
#> 34              260          0
#> 35              261          0
#> 36              775          0
#> 37              820          0
#> 38              821          0
#> 39              261          0
#> 40              529          0
#> 41              821          0
#> 42              525          0
#> 43              691          0
#> 44              397          0
#> 45              691          0
#>                                                                                                                                  message
#> 3                                                                                   Reception type (114) does not match serve type (114)
#> 4                                            Front-row player made an attack from a back-row zone (legal, but possibly a scouting error)
#> 5                                                                                                             Block by a back-row player
#> 21                                                                  Repeated row with same skill and evaluation_code for the same player
#> 22                                                                                                Consecutive actions by the same player
#> 23                                                                                                Consecutive actions by the same player
#> 24                                                                                                Consecutive actions by the same player
#> 25                                                                                                Consecutive actions by the same player
#> 26                                                                                                Consecutive actions by the same player
#> 27                                                                                                Consecutive actions by the same player
#> 28                                                                                                Consecutive actions by the same player
#> 29                                                                                                Consecutive actions by the same player
#> 30                                                                                                Consecutive actions by the same player
#> 31                                                                                     Rally does not include a winning or losing action
#> 32                                                                                     Rally does not include a winning or losing action
#> 33                                                     Point awarded to incorrect team following error (or "error" evaluation incorrect)
#> 34                                                            Point awarded to incorrect team (or "Winning attack" evaluation incorrect)
#> 35                                                             Point awarded to incorrect team (or "Winning block" evaluation incorrect)
#> 36                                                            Point awarded to incorrect team (or "Winning attack" evaluation incorrect)
#> 37                                                            Point awarded to incorrect team (or "Winning attack" evaluation incorrect)
#> 38                                                             Point awarded to incorrect team (or "Winning block" evaluation incorrect)
#> 39 Point assigned to incorrect team or scores incorrect (point was won by UTSSU Men's but score was incremented for Canberra Heat Men's)
#> 40 Point assigned to incorrect team or scores incorrect (point was won by Canberra Heat Men's but score was incremented for UTSSU Men's)
#> 41 Point assigned to incorrect team or scores incorrect (point was won by UTSSU Men's but score was incremented for Canberra Heat Men's)
#> 42                                        Scores do not follow proper sequence (note that the error may be in the point before this one)
#> 43                                                                                            Home team rotation has changed incorrectly
#> 44                                                                                        Visiting team rotation has changed incorrectly
#> 45                                                                                        Visiting team rotation has changed incorrectly
#>                                                                                                                                                                                                                                                                                                                                                                                                      file_line
#> 3                                    E~{"PlayerGuid":"EE1E33FC-2BE7-4374-9D0B-C8BF9B3D1B0A-29139-00009C8C6E6243F6","videoDuration":0,"EventString":"Pass","BallEndString":"0, 0|0","EventGrade":1,"TimeStamp":"2017-09-23T00:25:03.000Z","ErrorType":0,"OppositionScore":8,"SubEvent2":0,"TeamScore":8,"SubEvent":0,"EventId":"620","EventType":2,"Row":4,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 4  E~{"PlayerGuid":"EE1E33FC-2BE7-4374-9D0B-C8BF9B3D1B0A-29139-00009C8C6E6243F6","videoDuration":0,"EventString":"Spike","BallEndString":"  -1.67,  -1.67","EventGrade":3,"TimeStamp":"2017-09-23T01:42:24.000Z","ErrorType":0,"OppositionScore":15,"SubEvent2":0,"TeamScore":23,"SubEvent":4113,"EventId":"5430","EventType":4,"Row":1,"BallMidString":"  87.08, 102.08","BallStartString":" 102.92, 143.75"}
#> 5                                E~{"PlayerGuid":"3DB5973B-0078-455A-AAAD-6D0A6023571F-29139-00009C681BD6229E","videoDuration":0,"EventString":"Block","BallEndString":"0, 0|0","EventGrade":3,"TimeStamp":"2017-09-23T01:42:41.000Z","ErrorType":0,"OppositionScore":15,"SubEvent2":0,"TeamScore":24,"SubEvent":0,"EventId":"5435","EventType":5,"Row":2,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 21                             E~{"PlayerGuid":"4215B3CA-3F96-4463-A5C1-1412D8BEDFD5-88820-0001897E5C205DD9","videoDuration":0,"EventString":"Defense","BallEndString":"0, 0|0","EventGrade":1,"TimeStamp":"2017-09-23T00:35:10.000Z","ErrorType":0,"OppositionScore":18,"SubEvent2":0,"TeamScore":18,"SubEvent":0,"EventId":"1350","EventType":6,"Row":5,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 22       E~{"PlayerGuid":"4215B3CA-3F96-4463-A5C1-1412D8BEDFD5-88820-0001897E5C205DD9","videoDuration":0,"EventString":"Spike","BallEndString":"  66.67,  88.33","EventGrade":1,"TimeStamp":"2017-09-23T00:17:52.000Z","ErrorType":0,"OppositionScore":1,"SubEvent2":0,"TeamScore":1,"SubEvent":4,"EventId":"110","EventType":4,"Row":1,"BallMidString":"  25.00,  96.67","BallStartString":"   7.92,  87.92"}
#> 23    E~{"PlayerGuid":"EE1E33FC-2BE7-4374-9D0B-C8BF9B3D1B0A-29139-00009C8C6E6243F6","videoDuration":0,"EventString":"Spike","BallEndString":"  54.58, 133.75","EventGrade":1,"TimeStamp":"2017-09-23T00:19:32.000Z","ErrorType":0,"OppositionScore":3,"SubEvent2":0,"TeamScore":3,"SubEvent":4115,"EventId":"230","EventType":4,"Row":1,"BallMidString":"  87.08,  95.42","BallStartString":" 102.92, 113.75"}
#> 24     E~{"PlayerGuid":"3DB5973B-0078-455A-AAAD-6D0A6023571F-29139-00009C681BD6229E","videoDuration":0,"EventString":"Spike","BallEndString":"  87.92,  39.17","EventGrade":1,"TimeStamp":"2017-09-23T00:24:19.000Z","ErrorType":0,"OppositionScore":7,"SubEvent2":0,"TeamScore":8,"SubEvent":111,"EventId":"530","EventType":4,"Row":4,"BallMidString":"   0.00,   0.00","BallStartString":"   5.42, 116.25"}
#> 25 E~{"PlayerGuid":"3DB5973B-0078-455A-AAAD-6D0A6023571F-29139-00009C681BD6229E","videoDuration":0,"EventString":"Spike","BallEndString":"  22.50,  78.33","EventGrade":1,"TimeStamp":"2017-09-23T00:36:06.000Z","ErrorType":0,"OppositionScore":20,"SubEvent2":0,"TeamScore":18,"SubEvent":1111,"EventId":"1410","EventType":4,"Row":4,"BallMidString":"   0.00,   0.00","BallStartString":"   7.92, 114.17"}
#> 26    E~{"PlayerGuid":"EE1E33FC-2BE7-4374-9D0B-C8BF9B3D1B0A-29139-00009C8C6E6243F6","videoDuration":0,"EventString":"Spike","BallEndString":"  35.00, 180.42","EventGrade":3,"TimeStamp":"2017-09-23T00:53:00.000Z","ErrorType":0,"OppositionScore":8,"SubEvent2":0,"TeamScore":9,"SubEvent":111,"EventId":"2350","EventType":4,"Row":6,"BallMidString":"   0.00,   0.00","BallStartString":"  96.67,  87.08"}
#> 27 E~{"PlayerGuid":"3DB5973B-0078-455A-AAAD-6D0A6023571F-29139-00009C681BD6229E","videoDuration":0,"EventString":"Spike","BallEndString":" 110.83,  61.25","EventGrade":3,"TimeStamp":"2017-09-23T01:18:33.000Z","ErrorType":0,"OppositionScore":29,"SubEvent2":0,"TeamScore":28,"SubEvent":4111,"EventId":"4050","EventType":4,"Row":2,"BallMidString":"  95.00, 101.67","BallStartString":" 107.08,  83.33"}
#> 28    E~{"PlayerGuid":"37326FEF-7872-4D50-98E0-00B6BDFF6CE4-88820-00018987192EB04E","videoDuration":0,"EventString":"Spike","BallEndString":"  65.42, 139.58","EventGrade":1,"TimeStamp":"2017-09-23T01:19:06.000Z","ErrorType":0,"OppositionScore":29,"SubEvent2":0,"TeamScore":29,"SubEvent":4,"EventId":"4090","EventType":4,"Row":3,"BallMidString":"  17.92,  97.08","BallStartString":"   1.67, 115.83"}
#> 29   E~{"PlayerGuid":"4215B3CA-3F96-4463-A5C1-1412D8BEDFD5-88820-0001897E5C205DD9","videoDuration":0,"EventString":"Spike","BallEndString":" -12.08, 140.00","EventGrade":0,"TimeStamp":"2017-09-23T01:31:49.000Z","ErrorType":1,"OppositionScore":6,"SubEvent2":0,"TeamScore":11,"SubEvent":111,"EventId":"4800","EventType":4,"Row":5,"BallMidString":"   0.00,   0.00","BallStartString":"  95.42,  92.50"}
#> 30  E~{"PlayerGuid":"EE1E33FC-2BE7-4374-9D0B-C8BF9B3D1B0A-29139-00009C8C6E6243F6","videoDuration":0,"EventString":"Spike","BallEndString":"  73.33,  42.50","EventGrade":3,"TimeStamp":"2017-09-23T01:41:29.000Z","ErrorType":0,"OppositionScore":14,"SubEvent2":0,"TeamScore":22,"SubEvent":111,"EventId":"5390","EventType":4,"Row":6,"BallMidString":"   0.00,   0.00","BallStartString":"  -1.25, 107.50"}
#> 31   E~{"PlayerGuid":"511AE5E6-BE3B-40C8-94BB-313968399497-88820-0001896A31B035C3","videoDuration":0,"EventString":"Spike","BallEndString":"  35.00,  70.83","EventGrade":0,"TimeStamp":"2017-09-23T00:27:57.000Z","ErrorType":2,"OppositionScore":10,"SubEvent2":0,"TeamScore":13,"SubEvent":113,"EventId":"890","EventType":4,"Row":6,"BallMidString":"  25.00, 104.58","BallStartString":"   9.58,  58.75"}
#> 32  E~{"PlayerGuid":"CB72B343-0910-42EF-8750-AB96A016C503-88820-000189A294528089","videoDuration":0,"EventString":"Spike","BallEndString":"  62.50,  88.75","EventGrade":0,"TimeStamp":"2017-09-23T01:42:42.000Z","ErrorType":2,"OppositionScore":15,"SubEvent2":0,"TeamScore":24,"SubEvent":112,"EventId":"5460","EventType":4,"Row":1,"BallMidString":"  41.67, 102.50","BallStartString":"  32.08,  82.92"}
#> 33                               E~{"PlayerGuid":"3DB5973B-0078-455A-AAAD-6D0A6023571F-29139-00009C681BD6229E","videoDuration":0,"EventString":"Serve","BallEndString":"0, 0|0","EventGrade":0,"TimeStamp":"2017-09-23T00:37:38.000Z","ErrorType":0,"OppositionScore":21,"SubEvent2":0,"TeamScore":19,"SubEvent":0,"EventId":"1460","EventType":1,"Row":5,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 34  E~{"PlayerGuid":"EE1E33FC-2BE7-4374-9D0B-C8BF9B3D1B0A-29139-00009C8C6E6243F6","videoDuration":0,"EventString":"Spike","BallEndString":"  69.17,  29.58","EventGrade":3,"TimeStamp":"2017-09-23T00:27:32.000Z","ErrorType":0,"OppositionScore":10,"SubEvent2":0,"TeamScore":12,"SubEvent":3111,"EventId":"850","EventType":4,"Row":6,"BallMidString":"   0.00,   0.00","BallStartString":"   6.67, 111.67"}
#> 35                                E~{"PlayerGuid":"EE1E33FC-2BE7-4374-9D0B-C8BF9B3D1B0A-29139-00009C8C6E6243F6","videoDuration":0,"EventString":"Block","BallEndString":"0, 0|0","EventGrade":3,"TimeStamp":"2017-09-23T00:27:56.000Z","ErrorType":0,"OppositionScore":10,"SubEvent2":0,"TeamScore":13,"SubEvent":0,"EventId":"860","EventType":5,"Row":6,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 36     E~{"PlayerGuid":"511AE5E6-BE3B-40C8-94BB-313968399497-88820-0001896A31B035C3","videoDuration":0,"EventString":"Spike","BallEndString":"  17.92, 119.58","EventGrade":3,"TimeStamp":"2017-09-23T01:34:49.000Z","ErrorType":0,"OppositionScore":8,"SubEvent2":0,"TeamScore":15,"SubEvent":4,"EventId":"4980","EventType":4,"Row":1,"BallMidString":"  23.33, 100.83","BallStartString":"  -3.33,  86.25"}
#> 37 E~{"PlayerGuid":"EE1E33FC-2BE7-4374-9D0B-C8BF9B3D1B0A-29139-00009C8C6E6243F6","videoDuration":0,"EventString":"Spike","BallEndString":"  -1.67,  -1.67","EventGrade":3,"TimeStamp":"2017-09-23T01:42:24.000Z","ErrorType":0,"OppositionScore":15,"SubEvent2":0,"TeamScore":23,"SubEvent":4113,"EventId":"5430","EventType":4,"Row":1,"BallMidString":"  87.08, 102.08","BallStartString":" 102.92, 143.75"}
#> 38                               E~{"PlayerGuid":"3DB5973B-0078-455A-AAAD-6D0A6023571F-29139-00009C681BD6229E","videoDuration":0,"EventString":"Block","BallEndString":"0, 0|0","EventGrade":3,"TimeStamp":"2017-09-23T01:42:41.000Z","ErrorType":0,"OppositionScore":15,"SubEvent2":0,"TeamScore":24,"SubEvent":0,"EventId":"5435","EventType":5,"Row":2,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 39                                E~{"PlayerGuid":"EE1E33FC-2BE7-4374-9D0B-C8BF9B3D1B0A-29139-00009C8C6E6243F6","videoDuration":0,"EventString":"Block","BallEndString":"0, 0|0","EventGrade":3,"TimeStamp":"2017-09-23T00:27:56.000Z","ErrorType":0,"OppositionScore":10,"SubEvent2":0,"TeamScore":13,"SubEvent":0,"EventId":"860","EventType":5,"Row":6,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 40                             E~{"PlayerGuid":"8D6CE526-CD36-45C3-8F89-47A6AC7F41B4-88820-000189786DF4CC1D","videoDuration":0,"EventString":"Defense","BallEndString":"0, 0|0","EventGrade":0,"TimeStamp":"2017-09-23T01:05:41.000Z","ErrorType":0,"OppositionScore":17,"SubEvent2":0,"TeamScore":18,"SubEvent":0,"EventId":"3030","EventType":6,"Row":1,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 41                               E~{"PlayerGuid":"3DB5973B-0078-455A-AAAD-6D0A6023571F-29139-00009C681BD6229E","videoDuration":0,"EventString":"Block","BallEndString":"0, 0|0","EventGrade":3,"TimeStamp":"2017-09-23T01:42:41.000Z","ErrorType":0,"OppositionScore":15,"SubEvent2":0,"TeamScore":24,"SubEvent":0,"EventId":"5435","EventType":5,"Row":2,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 42                               E~{"PlayerGuid":"AC9507BF-2C3A-4D70-9745-4337C44F4514-29139-00009C46F9321D2B","videoDuration":0,"EventString":"Serve","BallEndString":"0, 0|0","EventGrade":2,"TimeStamp":"2017-09-23T01:05:16.000Z","ErrorType":0,"OppositionScore":17,"SubEvent2":0,"TeamScore":17,"SubEvent":0,"EventId":"2990","EventType":1,"Row":1,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 43                                 E~{"PlayerGuid":"8D6CE526-CD36-45C3-8F89-47A6AC7F41B4-88820-000189786DF4CC1D","videoDuration":0,"EventString":"Serve","BallEndString":"0, 0|0","EventGrade":1,"TimeStamp":"2017-09-23T01:23:04.000Z","ErrorType":0,"OppositionScore":0,"SubEvent2":0,"TeamScore":0,"SubEvent":0,"EventId":"4140","EventType":1,"Row":0,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 44                                 E~{"PlayerGuid":"AC9507BF-2C3A-4D70-9745-4337C44F4514-29139-00009C46F9321D2B","videoDuration":0,"EventString":"Serve","BallEndString":"0, 0|0","EventGrade":1,"TimeStamp":"2017-09-23T00:44:57.000Z","ErrorType":0,"OppositionScore":0,"SubEvent2":0,"TeamScore":0,"SubEvent":0,"EventId":"1710","EventType":1,"Row":1,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}
#> 45                                 E~{"PlayerGuid":"8D6CE526-CD36-45C3-8F89-47A6AC7F41B4-88820-000189786DF4CC1D","videoDuration":0,"EventString":"Serve","BallEndString":"0, 0|0","EventGrade":1,"TimeStamp":"2017-09-23T01:23:04.000Z","ErrorType":0,"OppositionScore":0,"SubEvent2":0,"TeamScore":0,"SubEvent":0,"EventId":"4140","EventType":1,"Row":0,"BallMidString":"0, 0|0","BallStartString":"0, 0|0"}